TASK-44896 Review News extension mecanism and define News activity actions (#122)
Prior to this change, the Activity composer extension defined by news is inserted sometimes late in render time, thus we defined a portlet module that will be injected in a shared AddonContainer in all pages to insert news extension in extensionRegistry to be used when needed
