+++ webapp/src/main/webapp/WEB-INF/portlet.xml
@@ -42,4 +42,24 @@
         </portlet-info>
     </portlet>
 
+    <portlet>
+        <portlet-name>NewsExtensions</portlet-name>
+        <display-name xml:lang="EN">News UI Extensions</display-name>
+        <portlet-class>org.exoplatform.commons.api.portlet.GenericDispatchedViewPortlet</portlet-class>
+        <init-param>
+            <name>portlet-view-dispatched-file-path</name>
+            <value>/news-extensions/newsExtensions.html</value>
+        </init-param>
+        <expiration-cache>-1</expiration-cache>
+        <cache-scope>PUBLIC</cache-scope>
+        <supports>
+            <mime-type>text/html</mime-type>
+        </supports>
+        <supported-locale>en</supported-locale>
+        <portlet-info>
+            <title>News UI Extensions</title>
+            <keywords>News UI Extensions</keywords>
+        </portlet-info>
+    </portlet>
+
 </portlet-app>
