--- webapp/src/main/webapp/groovy/news/webui/composer/UINewsActivityComposer.gtmpl
@@ -1,17 +1,0 @@
-<%
-def requestContext = _ctx.getRequestContext();
-def jsManager = requestContext.getJavascriptManager();
-
-if (uicomponent.isEnabled()) {
-  def space = uicomponent.getCurrentSpace();
-  def spaceId = space?.getId();
-  def labels = """ {
-    writeLinkLabel: '${_ctx.appRes("news.composer.write")}'
-  } """
-
-  jsManager.require("SHARED/newsComposerPlugin","newsComposerPlugin")
-         .require("SHARED/newsComposerIntegration","newsComposerIntegration")
-         .addScripts("newsComposerIntegration.init($spaceId, $labels) ;");
-}
-
-%>
