+++ webapp/src/main/webapp/news-extensions/main.js
@@ -1,0 +1,13 @@
+import * as  newsServices from '../services/newsServices';
+
+if (!Vue.prototype.$newsServices) {
+  window.Object.defineProperty(Vue.prototype, '$newsServices', {
+    value: newsServices,
+  });
+}
+
+import {initExtensions} from './extensions.js';
+
+export function init() {
+  initExtensions();
+}
