+++ webapp/src/main/webapp/news-extensions/newsExtensions.html
@@ -1,0 +1,3 @@
+<script>
+  require(['PORTLET/news/NewsExtensions'], extensions => extensions.init());
+</script>
